# openapi_schema.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Cinema ERP - Firebase Simples",
                 "description":  "Sistema funcional com PostgreSQL e Firebase",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/locations/search":  {
                                                   "get":  {
                                                               "summary":  "Search Locations",
                                                               "description":  "Buscar locaÃ§Ãµes com filtros",
                                                               "operationId":  "search_locations_api_v1_locations_search_get",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "q",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Q"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "city",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "City"
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "status",
                                                                                      "in":  "query",
                                                                                      "required":  false,
                                                                                      "schema":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Status"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {

                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           },
                                                   "post":  {
                                                                "summary":  "Search Locations",
                                                                "description":  "Busca avanÃ§ada de locaÃ§Ãµes",
                                                                "operationId":  "search_locations_api_v1_locations_search_post",
                                                                "requestBody":  {
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "object",
                                                                                                                                            "additionalProperties":  true,
                                                                                                                                            "title":  "Search Data"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Successful Response",
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {

                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          },
                                                                                  "422":  {
                                                                                              "description":  "Validation Error",
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                               },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}